Entity IDs need @GeneratedValue if ID not set manually
Change the type of User.id from long to Long, so that it is default
initialized to null instead of 0. 0 would be a valid ID, having it
default to null exposes the behavior I am interested in.

I wasn't sure if I needed @GeneratedValue on User's `id` property. I
didn't know whether JPA/Hibernate just needed it for generating a schema
(which I don't want it to do; I define the schema by hand with
Liquibase) or if it needed it even if I just use JPA for OR mapping.
Turns out it does need it! The test fails with the following error
(which the test unfortunately does not express clearly):

org.springframework.orm.jpa.JpaSystemException: ids for this class must
be manually assigned before calling save(): name.cgt.demo.learnjpa.User;
nested exception is org.hibernate.id.IdentifierGenerationException: ids
for this class must be manually assigned before calling save():
name.cgt.demo.learnjpa.User